WIP: toward coreProposal for vat-network
 - no vat-network in prod
   - only in boot-sim (hm... not good enough for devnet)
 - coreProposals supports installations but not raw vats,
   so we're doing static config for now
 - split registerNetworkProtocols into client, chain cases
