+++ packages/vats/src/core/boot-sim.js
@@ -4,6 +4,7 @@ import { makeBootstrap } from './lib-boot.js';
 import * as basicBehaviorsPlus from './basic-behaviors.js';
 import * as chainBehaviorsPlus from './chain-behaviors.js';
 import * as simBehaviorsPlus from './sim-behaviors.js';
+import * as startNetPlus from './startNetIBC.js';
 import * as utils from './utils.js';
 
 const { BASIC_BOOTSTRAP_PERMITS: _b, ...basicBehaviors } = basicBehaviorsPlus;
@@ -13,13 +14,20 @@ const {
   ...chainBehaviors
 } = chainBehaviorsPlus;
 const { SIM_CHAIN_BOOTSTRAP_PERMITS, ...simBehaviors } = simBehaviorsPlus;
+const { NET_MANIFEST, getNetIBCManifest: _g, ...startNet } = startNetPlus;
 
 export const MANIFEST = {
   ...SHARED_CHAIN_BOOTSTRAP_MANIFEST,
   ...SIM_CHAIN_BOOTSTRAP_PERMITS,
+  ...NET_MANIFEST,
 };
 
-const behaviors = { ...basicBehaviors, ...chainBehaviors, ...simBehaviors };
+const behaviors = {
+  ...basicBehaviors,
+  ...chainBehaviors,
+  ...simBehaviors,
+  ...startNet,
+};
 
 const modules = harden({ behaviors: { ...behaviors }, utils: { ...utils } });
