+++ packages/vats/src/core/types.js
@@ -256,8 +256,8 @@
  * @typedef {ERef<ReturnType<import('../vat-provisioning.js').buildRootObject>>} ProvisioningVat
  * @typedef {ERef<ReturnType<import('../vat-mints.js').buildRootObject>>} MintsVat
  * @typedef {ERef<ReturnType<import('../vat-priceAuthority.js').buildRootObject>>} PriceAuthorityVat
- * @typedef {ERef<ReturnType<import('../vat-network.js').buildRootObject>>} NetworkVat
- * @typedef {ERef<ReturnType<import('../vat-ibc.js').buildRootObject>>} IBCVat
+ * @typedef {ReturnType<import('../vat-network.js').buildRootObject>} NetworkVat
+ * @typedef {ReturnType<import('../vat-ibc.js').buildRootObject>} IBCVat
  * @typedef { import('@agoric/zoe/tools/priceAuthorityRegistry').PriceAuthorityRegistryAdmin } PriceAuthorityRegistryAdmin
  */
